Implement TWI/I2C network stack for the STM32F3 family of chips.
Send and receive side both work.

I believe the STM32F3 uses the same "V2" I2C peripheral shared by
the STM32F0 and the STM32Lx so this should, I hope, be the last
time we ever have to port this code to a new I2C peripheral.

Currently this only supports a single I2C peripheral but the code
should be structured such that it's not too hard to add
multi-peripheral support in the future. I hope.