Improve duktig control program to cycle between off/low/high
After seeing a youtube video of a child playing with the authentic Duktig, I
realized the light doesn't just turn off and on: it cycles between off/low/high.

Unfortunately the hardware is already locked and one of the two LED control
inputs is not on an stm32 output pin that is attached to a timer's PWM output.
This PR instead uses software PWM by driving the LED pin using a clock interrupt.