Change GPIO pin macros to real types instead of preprocessor magic
gpio_set, gpio_clr, etc, used to be comma-separated lists of config data meant
to be passed into functions that take the same number of arguments. This was
pointless preprocessor magic that made it impossible to store a pin definition
dynamically for later use. The PINUSE and PINREF macros then came about to
satisfy that use case.

There is not much point to this; this PR just makes all the gpio functions take
proper structures, so storing a pin for later use is no longer a special case.