Remove now-deprecated rocket0 and rocket1 apps.
The rocket is dead. Long live the rocket!

The rocket0 and rocket1 apps harken back to the original version of the rocket
when we had two separate CPUs that each controlled a local set of displays via
an EPB, and passed high level protocol messages to each other via TWI.

This architecture has since been deprecated. On the modern rocket, the entire
app runs in a single CPU. Individual display frames are now communicated over
TWI instead of EPB.