Fix race condition in linereader: do not pop unconsumed bytes after last EOL
The bug is not normally seen in low-load situations because the handler runs so
quickly that it runs after each character, so there's never any data meant to be
leftover. But in real app that has higher load, the handler can run late after
multiple characters have been received and the bug manifests.