Only configure UART GPIO pins on-demand
Before this change, UART initialization would also initialize the RX and TX
pins, activating the alternate function to bind them to the UART. However, some
applications want to use only half the UART (e.g. only transmit, or only
receive) and want to use the other pin as a GPIO. This can be accomplished
simply by not configuring the pin in question as a UART pin.

This change does not configure the RX or TX pins until an RX or TX is
attempted. This way an application that wants to use the TX pin as a GPIO can do
so simply by not transmitting.