Remove all stm32 vendor libraries from repo; replace with submodules
It's long bothered me that we had hundreds of megs of external code into our
repo, but it seemed worthwhile so that the repo was relatively self-contained;
downloading the stm32 HAL libraries was a difficult and manual process that
could only be done by creating an account on the ST web site.

I recently discovered that ST now publishes their software on github, hooray!
This commit deletes all their software from our repo and references them using
submodules instead.