address dropped bytes
i had a theory that dropped bytes were caused by an asynchronous buffer swap that was not well aligned to the arrival of the next byte.  i added a loop to spin until the next byte, which seems to fix the problem. this probably only needs to be done when the trigger for this action was NOT the idle interrupt, but this code does it also in this case.